requirements.txt
python-telegram-bot==20.6
yfinance
pandas
ta
